crime report verification and assignment to police station is done

+++ src/main/java/com/app/epolice/controller/CrimeReportController.java
@@ -5,6 +5,7 @@ import com.app.epolice.service.CrimeReportService;
 import com.app.epolice.util.FileUpload;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.util.StringUtils;
@@ -20,7 +21,9 @@ import java.util.List;
 @RestController
 @RequestMapping("/crime-reports")
 public class CrimeReportController {
-    private static final Logger LOG = LogManager.getLogger(UserController.class);
+    private static final Logger LOG = LogManager.getLogger(CrimeReportController.class);
+    private static final String token = "40dc498b-e837-4fa9-8e53-c1d51e01af15";
+
 
     /**
      * Initializing the objects
@@ -31,12 +34,40 @@ public class CrimeReportController {
     }
 
     /**
+     * Authorizing the token
+     *
+     * @param token
+     * @return
+     * @Author "Kamran"
+     */
+    public boolean authorization(String token) {
+        LOG.info("Authorizing the user ");
+        return CrimeReportController.token.equals(token);
+    }
+
+    /**
+     * if the user is un-authorized
+     *
+     * @return
+     * @Author "Kamran"
+     */
+    public ResponseEntity<Object> unAuthorizeUser() {
+        LOG.info("Unauthorized user is trying to get access");
+        return new ResponseEntity<>("Kindly do the authorization first", HttpStatus.UNAUTHORIZED);
+    }
+
+    /**
      * Showing all the crimeReports
      * @return
      */
     @GetMapping("/list")
-    public ResponseEntity<Object> listOfCrimeReports(){
-        return crimeReportService.listAllCrimeReports();
+    public ResponseEntity<Object> listOfCrimeReports(@RequestHeader("Authorization") String token){
+        if (authorization(token)) {
+            LOG.info("listing all the crime reports");
+            return crimeReportService.listAllCrimeReports();
+        } else {
+            return unAuthorizeUser();
+        }
     }
 
     /**
@@ -45,8 +76,13 @@ public class CrimeReportController {
      * @return
      */
     @PostMapping("/add_multiple_reports")
-    public ResponseEntity<Object> addListOfCrimeReports(@RequestBody List<CrimeReport> crimeReport){
-        return crimeReportService.addMultipleCrimeReports(crimeReport);
+    public ResponseEntity<Object> addListOfCrimeReports(@RequestHeader("Authorization") String token, @RequestBody List<CrimeReport> crimeReport){
+        if (authorization(token)) {
+            LOG.info("adding the crime reports");
+            return crimeReportService.addMultipleCrimeReports(crimeReport);
+        } else {
+            return unAuthorizeUser();
+        }
     }
 
     /**
@@ -57,8 +93,13 @@ public class CrimeReportController {
      * @throws IOException
      */
     @PostMapping("/upload_singe_report")
-    public ResponseEntity<Object> uploadReport( CrimeReport report,@RequestParam("files") MultipartFile[] file) throws IOException {
-        return crimeReportService.addSingleCrimeReport(report,file);
+    public ResponseEntity<Object> uploadReport(@RequestHeader("Authorization") String token, CrimeReport report,@RequestParam("files") MultipartFile[] file) throws IOException {
+        if (authorization(token)) {
+            LOG.info("adding single crime report");
+            return crimeReportService.addSingleCrimeReport(report,file);
+        } else {
+            return unAuthorizeUser();
+        }
     }
 
     /**
@@ -68,7 +109,7 @@ public class CrimeReportController {
      * @throws IOException
      */
     @RequestMapping(value = "/upload", method = RequestMethod.POST, consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public String uploadFile(@RequestParam("file") MultipartFile file) throws IOException {
+    public String uploadFile(@RequestHeader("Authorization") String token, @RequestParam("file") MultipartFile file) throws IOException {
         String reportPictureName = StringUtils.cleanPath(file.getOriginalFilename());
         String uploadDir = "F:\\Development\\E-Police Project\\Images\\" + 1;
         FileUpload.saveFile(uploadDir,reportPictureName, file);
@@ -81,8 +122,13 @@ public class CrimeReportController {
      * @return
      */
     @PutMapping("/update")
-    public ResponseEntity<Object> updateCrimeReport(@RequestBody CrimeReport crimeReport){
-        return crimeReportService.updateCrimeReport(crimeReport);
+    public ResponseEntity<Object> updateCrimeReport(@RequestHeader("Authorization") String token, @RequestBody CrimeReport crimeReport){
+        if (authorization(token)) {
+            LOG.info("updating single crime report");
+            return crimeReportService.updateCrimeReport(crimeReport);
+        } else {
+            return unAuthorizeUser();
+        }
     }
 
     /**
@@ -91,8 +137,13 @@ public class CrimeReportController {
      * @return
      */
     @DeleteMapping("/delete")
-    public ResponseEntity<Object> deleteCrimeReport(@RequestBody List<CrimeReport> crimeReportList){
-        return crimeReportService.deleteCrimeReport(crimeReportList);
+    public ResponseEntity<Object> deleteCrimeReport(@RequestHeader("Authorization") String token, @RequestBody List<CrimeReport> crimeReportList){
+        if (authorization(token)) {
+            LOG.info("deleting multiple crime reports");
+            return crimeReportService.deleteCrimeReport(crimeReportList);
+        } else {
+            return unAuthorizeUser();
+        }
     }
 
     /**
@@ -101,7 +152,12 @@ public class CrimeReportController {
      * @return
      */
     @PostMapping("/verify")
-    public ResponseEntity<Object> verifyCrimeReport(@RequestHeader String status,@RequestHeader long crimeReportId) throws ParseException {
-        return crimeReportService.verifyReport(status,crimeReportId);
+    public ResponseEntity<Object> verifyCrimeReport(@RequestHeader("Authorization") String token, @RequestHeader String status,@RequestHeader long crimeReportId,@RequestHeader long policeStationId) throws ParseException {
+        if (authorization(token)) {
+            LOG.info("verifying and assigning crime reports to a specific police station");
+            return crimeReportService.verifyReport(status,crimeReportId,policeStationId);
+        } else {
+            return unAuthorizeUser();
+        }
     }
 }
