+++ src/main/java/com/app/epolice/controller/CriminalController.java
@@ -4,6 +4,7 @@ import com.app.epolice.model.entity.crime.Criminal;
 import com.app.epolice.service.CriminalService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
@@ -14,11 +15,35 @@ import java.util.List;
 @RestController
 @RequestMapping("/criminal")
 public class CriminalController {
-    private static final Logger LOG = LogManager.getLogger(UserController.class);
-    CriminalService criminalService;
+    private static final Logger LOG = LogManager.getLogger(CriminalController.class);
+    private static final String token = "40dc498b-e837-4fa9-8e53-c1d51e01af15";
 
+    CriminalService criminalService;
     public CriminalController(CriminalService criminalService) {
         this.criminalService = criminalService;
+    }
+
+    /**
+     * Authorizing the token
+     *
+     * @param token
+     * @return
+     * @Author "Kamran"
+     */
+    public boolean authorization(String token) {
+        LOG.info("Authorizing the user ");
+        return CriminalController.token.equals(token);
+    }
+
+    /**
+     * if the user is un-authorized
+     *
+     * @return
+     * @Author "Kamran"
+     */
+    public ResponseEntity<Object> unAuthorizeUser() {
+        LOG.info("Unauthorized user is trying to get access");
+        return new ResponseEntity<>("Kindly do the authorization first", HttpStatus.UNAUTHORIZED);
     }
 
     /**
@@ -26,8 +51,13 @@ public class CriminalController {
      * @return
      */
     @GetMapping("/list")
-    public ResponseEntity<Object> listOfCriminals(){
-        return criminalService.listAllCriminals();
+    public ResponseEntity<Object> listOfCriminals(@RequestHeader("Authorization") String token){
+        if (authorization(token)) {
+            LOG.info("Listing all the criminals");
+            return criminalService.listAllCriminals();
+        } else {
+            return unAuthorizeUser();
+        }
     }
 
     /**
@@ -36,8 +66,13 @@ public class CriminalController {
      * @return
      */
     @PostMapping("/add")
-    public ResponseEntity<Object> addCriminal(@RequestBody List<Criminal> criminal){
-        return criminalService.addNewCriminals(criminal);
+    public ResponseEntity<Object> addCriminal(@RequestHeader("Authorization") String token, @RequestBody List<Criminal> criminal){
+        if (authorization(token)) {
+            LOG.info("adding the criminals");
+            return criminalService.addNewCriminals(criminal);
+        } else {
+            return unAuthorizeUser();
+        }
     }
 
     /**
@@ -46,8 +81,13 @@ public class CriminalController {
      * @return
      */
     @PutMapping("/update")
-    public ResponseEntity<Object> updateCriminal(@RequestBody Criminal criminal){
-        return criminalService.updateCriminal(criminal);
+    public ResponseEntity<Object> updateCriminal(@RequestHeader("Authorization") String token, @RequestBody Criminal criminal){
+        if (authorization(token)) {
+            LOG.info("updating the criminals");
+            return criminalService.updateCriminal(criminal);
+        } else {
+            return unAuthorizeUser();
+        }
     }
 
     /**
@@ -56,8 +96,13 @@ public class CriminalController {
      * @return
      */
     @DeleteMapping("/delete")
-    public ResponseEntity<Object> deleteCriminal(@RequestBody List<Criminal> criminalList){
-        return criminalService.deleteCriminal(criminalList);
+    public ResponseEntity<Object> deleteCriminal(@RequestHeader("Authorization") String token, @RequestBody List<Criminal> criminalList){
+        if (authorization(token)) {
+            LOG.info("deleting the criminals");
+            return criminalService.deleteCriminal(criminalList);
+        } else {
+            return unAuthorizeUser();
+        }
     }
 
     /**
@@ -66,16 +111,22 @@ public class CriminalController {
      * @return
      */
     @GetMapping("/find-by-cnic")
-    public ResponseEntity<Object> findCriminalByCnic(@RequestHeader String cnic){
-        return criminalService.findCriminalByCnic(cnic);
+    public ResponseEntity<Object> findCriminalByCnic(@RequestHeader("Authorization") String token, @RequestHeader String cnic){
+        if (authorization(token)) {
+            LOG.info("find the criminals by cnic");
+            return criminalService.findCriminalByCnic(cnic);
+        } else {
+            return unAuthorizeUser();
+        }
     }
+
     /**
      * find criminals by their cnic having return type boolean
      * @param cnic
      * @return
      */
     @GetMapping("/check-criminal-record")
-    public boolean verifyPersonCriminalRecord(@RequestHeader String cnic){
+    public boolean verifyPersonCriminalRecord(@RequestHeader("Authorization") String token, @RequestHeader String cnic){
         return criminalService.verifyPersonCriminalRecord(cnic);
     }
 
@@ -84,7 +135,7 @@ public class CriminalController {
      * @return
      */
     @GetMapping("/check-currency-record")
-    public String checkFeignCurrencyMethod(){
+    public String checkFeignCurrencyMethod(@RequestHeader("Authorization") String token){
         return criminalService.checkFeignCurrencyMethod();
     }
 }
