+++ src/main/java/com/app/epolice/controller/InvestigationTeamController.java
@@ -4,6 +4,7 @@ import com.app.epolice.model.entity.policestation.InvestigationTeam;
 import com.app.epolice.service.InvestigationTeamService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
@@ -14,11 +15,35 @@ import java.util.List;
 @RestController
 @RequestMapping("/investigation-team")
 public class InvestigationTeamController {
-    private static final Logger LOG = LogManager.getLogger(UserController.class);
-    InvestigationTeamService investigationTeamService;
+    private static final Logger LOG = LogManager.getLogger(InvestigationTeamController.class);
+    private static final String token = "40dc498b-e837-4fa9-8e53-c1d51e01af15";
 
+    InvestigationTeamService investigationTeamService;
     public InvestigationTeamController(InvestigationTeamService investigationTeamService) {
         this.investigationTeamService = investigationTeamService;
+    }
+
+    /**
+     * Authorizing the token
+     *
+     * @param token
+     * @return
+     * @Author "Kamran"
+     */
+    public boolean authorization(String token) {
+        LOG.info("Authorizing the user ");
+        return InvestigationTeamController.token.equals(token);
+    }
+
+    /**
+     * if the user is un-authorized
+     *
+     * @return
+     * @Author "Kamran"
+     */
+    public ResponseEntity<Object> unAuthorizeUser() {
+        LOG.info("Unauthorized user is trying to get access");
+        return new ResponseEntity<>("Kindly do the authorization first", HttpStatus.UNAUTHORIZED);
     }
 
     /**
@@ -26,8 +51,13 @@ public class InvestigationTeamController {
      * @return
      */
     @GetMapping("/list")
-    public ResponseEntity<Object> listOfInvestigationTeams(){
-        return investigationTeamService.listAllInvestigationTeams();
+    public ResponseEntity<Object> listOfInvestigationTeams(@RequestHeader("Authorization") String token){
+        if (authorization(token)) {
+            LOG.info("Listing all the investigation Teams");
+            return investigationTeamService.listAllInvestigationTeams();
+        } else {
+            return unAuthorizeUser();
+        }
     }
 
     /**
@@ -36,8 +66,13 @@ public class InvestigationTeamController {
      * @return
      */
     @PostMapping("/add")
-    public ResponseEntity<Object> addInvestigationTeam(@RequestBody List<InvestigationTeam> investigationTeam){
-        return investigationTeamService.addNewInvestigationTeams(investigationTeam);
+    public ResponseEntity<Object> addInvestigationTeam(@RequestHeader("Authorization") String token, @RequestBody List<InvestigationTeam> investigationTeam){
+        if (authorization(token)) {
+            LOG.info("Adding the investigation Teams");
+            return investigationTeamService.addNewInvestigationTeams(investigationTeam);
+        } else {
+            return unAuthorizeUser();
+        }
     }
 
     /**
@@ -46,8 +81,13 @@ public class InvestigationTeamController {
      * @return
      */
     @PutMapping("/update")
-    public ResponseEntity<Object> updateInvestigationTeam(@RequestBody InvestigationTeam investigationTeam){
-        return investigationTeamService.updateInvestigationTeam(investigationTeam);
+    public ResponseEntity<Object> updateInvestigationTeam(@RequestHeader("Authorization") String token, @RequestBody InvestigationTeam investigationTeam){
+        if (authorization(token)) {
+            LOG.info("updating the investigation Teams");
+            return investigationTeamService.updateInvestigationTeam(investigationTeam);
+        } else {
+            return unAuthorizeUser();
+        }
     }
 
     /**
@@ -56,7 +96,12 @@ public class InvestigationTeamController {
      * @return
      */
     @DeleteMapping("/delete")
-    public ResponseEntity<Object> deleteInvestigationTeam(@RequestBody List<InvestigationTeam> investigationTeamList){
-        return investigationTeamService.deleteInvestigationTeam(investigationTeamList);
+    public ResponseEntity<Object> deleteInvestigationTeam(@RequestHeader("Authorization") String token, @RequestBody List<InvestigationTeam> investigationTeamList){
+        if (authorization(token)) {
+            LOG.info("deleting the investigation Teams");
+            return investigationTeamService.deleteInvestigationTeam(investigationTeamList);
+        } else {
+            return unAuthorizeUser();
+        }
     }
 }
